+++ MyApplication/library/src/main/java/com/a65apps/kalimullinilnazrafilovich/library/applicaiton/fragments/ContactDetailsFragment.java
@@ -1,48 +1,34 @@
 package com.a65apps.kalimullinilnazrafilovich.library.applicaiton.fragments;
 
-import android.annotation.SuppressLint;
 import android.app.Application;
 import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.CompoundButton;
-import android.widget.TextView;
-import android.widget.ToggleButton;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.fragment.app.FragmentTransaction;
 
-import com.a65apps.kalimullinilnazrafilovich.entities.BirthdayNotification;
 import com.a65apps.kalimullinilnazrafilovich.entities.ContactDetailsInfo;
+import com.a65apps.kalimullinilnazrafilovich.library.applicaiton.delegate.details.ButtonsDelegate;
+import com.a65apps.kalimullinilnazrafilovich.library.applicaiton.delegate.details.ViewsDetailsDelegate;
 import com.a65apps.kalimullinilnazrafilovich.library.applicaiton.di.interfaces.ContactDetailsContainer;
 import com.a65apps.kalimullinilnazrafilovich.library.applicaiton.di.interfaces.HasAppContainer;
 import com.a65apps.kalimullinilnazrafilovich.library.applicaiton.presenters.ContactDetailsPresenter;
 import com.a65apps.kalimullinilnazrafilovich.library.applicaiton.views.ContactDetailsView;
 import com.a65apps.kalimullinilnazrafilovich.myapplication.R;
-import com.a65apps.kalimullinilnazrafilovich.myapplication.R2;
 
-import java.text.SimpleDateFormat;
-import java.util.GregorianCalendar;
-import java.util.Locale;
 import java.util.Objects;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
 
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.Unbinder;
 import moxy.MvpAppCompatFragment;
 import moxy.presenter.InjectPresenter;
 import moxy.presenter.ProvidePresenter;
 
-public class ContactDetailsFragment extends MvpAppCompatFragment implements
-        CompoundButton.OnCheckedChangeListener,
-        ContactDetailsView {
+public class ContactDetailsFragment extends MvpAppCompatFragment implements ContactDetailsView {
 
     @InjectPresenter
     @NonNull
@@ -51,32 +37,10 @@ public class ContactDetailsFragment extends MvpAppCompatFragment implements
     @NonNull
     public Provider<ContactDetailsPresenter> contactDetailsPresenterProvider;
 
-    @BindView(R2.id.name)
-    TextView name;
-    @BindView(R2.id.address)
-    TextView address;
-    @BindView(R2.id.firstTelephoneNumber)
-    TextView telephoneNumber;
-    @BindView(R2.id.secondTelephoneNumber)
-    TextView telephoneNumber2;
-    @BindView(R2.id.firstEmail)
-    TextView email;
-    @BindView(R2.id.secondEmail)
-    TextView email2;
-    @BindView(R2.id.description)
-    TextView description;
-    @BindView(R2.id.DayOfBirth)
-    TextView dataOfBirth;
-    @BindView(R2.id.btn_show_contact_location)
-    Button btnLocationOnMap;
-    @BindView(R2.id.btnBirthdayReminder)
-    ToggleButton btnBirthdayNotification;
-
-    private Unbinder unbinder;
+    private ViewsDetailsDelegate viewsDetailsDelegate;
+    private ButtonsDelegate buttonsDelegate;
 
     private String id;
-    private ContactDetailsInfo contactDetailsInfo;
-    private BirthdayNotification birthdayNotification;
 
     @NonNull
     public static ContactDetailsFragment newInstance(@NonNull String id) {
@@ -112,13 +76,14 @@ public class ContactDetailsFragment extends MvpAppCompatFragment implements
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_contact_details, container, false);
-        unbinder = ButterKnife.bind(this, view);
-
         Objects.requireNonNull(getActivity()).setTitle(getString(R.string.title_toolbar_contact_details));
 
-        id = getArguments().getString("id");
+        viewsDetailsDelegate = new ViewsDetailsDelegate(view);
+        buttonsDelegate = new ButtonsDelegate(view, getActivity(), contactDetailsPresenter);
 
-        btnLocationOnMap.setOnClickListener(v -> openMapFragment());
+        buttonsDelegate.clickButtons();
+
+        id = getArguments().getString("id");
 
         return view;
     }
@@ -131,66 +96,14 @@ public class ContactDetailsFragment extends MvpAppCompatFragment implements
 
     @Override
     public void showContactDetail(@NonNull ContactDetailsInfo contactDetailsInfo) {
-        this.contactDetailsInfo = contactDetailsInfo;
-
-        if (name != null) {
-            name.setText(contactDetailsInfo.getName());
-            dataOfBirth.setText(parseDateToString(contactDetailsInfo.getDateOfBirth()));
-            address.setText(contactDetailsInfo.getLocation().getAddress());
-            telephoneNumber.setText(contactDetailsInfo.getTelephoneNumber());
-            telephoneNumber2.setText(contactDetailsInfo.getTelephoneNumber2());
-            email.setText(contactDetailsInfo.getEmail());
-            email2.setText(contactDetailsInfo.getEmail2());
-            description.setText(contactDetailsInfo.getDescription());
-
-            setStatusLocationBtn(contactDetailsInfo.getLocation().getAddress());
-            setStatusToggleButton(btnBirthdayNotification);
-        }
-        setStatusLocationBtn(contactDetailsInfo.getLocation().getAddress());
-        setStatusToggleButton(btnBirthdayNotification);
-    }
-
-    private void setStatusToggleButton(@NonNull ToggleButton toggleButton) {
-        birthdayNotification = contactDetailsPresenter.getActualStateBirthdayNotification(contactDetailsInfo);
-        toggleButton.setChecked(birthdayNotification.isNotificationWorkStatusBoolean());
-        toggleButton.setOnCheckedChangeListener(this);
-    }
-
-    @Override
-    public void onCheckedChanged(@NonNull CompoundButton buttonView, boolean isChecked) {
-        if (isChecked) {
-            birthdayNotification = contactDetailsPresenter.setNotification(contactDetailsInfo);
-        } else {
-            birthdayNotification = contactDetailsPresenter.removeNotification(contactDetailsInfo);
-        }
-    }
-
-    private void setStatusLocationBtn(@NonNull String address) {
-        if ("".equals(address)) {
-            btnLocationOnMap.setText(R.string.status_btn_location_add);
-        } else {
-            btnLocationOnMap.setText(R.string.status_btn_location_check);
-        }
-    }
-
-    private void openMapFragment() {
-        ContactMapFragment contactMapFragment = ContactMapFragment.newInstance(id);
-        FragmentTransaction fragmentTransaction = Objects.requireNonNull(getActivity()).
-                getSupportFragmentManager().beginTransaction();
-        fragmentTransaction.replace(R.id.content, contactMapFragment).addToBackStack(null).commit();
-    }
-
-    @NonNull
-    private String parseDateToString(@NonNull GregorianCalendar gregorianCalendar) {
-        @SuppressLint("SimpleDateFormat")
-        SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd", Locale.getDefault());
-        df.setCalendar(gregorianCalendar);
-        return df.format(gregorianCalendar.getTime());
+        viewsDetailsDelegate.setViews(contactDetailsInfo);
+        buttonsDelegate.setStatusButtons(contactDetailsInfo);
     }
 
     @Override
     public void onDestroyView() {
         super.onDestroyView();
-        unbinder.unbind();
+        viewsDetailsDelegate.unBind();
+        buttonsDelegate.unBind();
     }
 }
