+++ MyApplication/library/src/main/res/layout/fragment_contact_details.xml
@@ -4,7 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".fragments.ContactListFragment"
-    android:background="@android:color/white"
     android:orientation="vertical">
 
     <LinearLayout
