+++ MyApplication/library/src/main/res/values/strings.xml
@@ -14,8 +14,8 @@
     <string name="title_notification">Напоминание о дне рождения</string>
     <string name="text_notification">сегодня празднует день рождения</string>
     <string name="deny_permission_message">Пожалуйста предоставте разрешение на чтение контактов</string>
-    <string name="google_maps_key">google_key</string>
-    <string name="yandex_maps_key">yandex_key</string>
+    <string name="google_maps_key">yy</string>
+    <string name="yandex_maps_key">uu</string>
     <string name="title_toolbar_contact_details">Детали контактов</string>
     <string name="tittle_toolbar_contact_list">Список контактов</string>
     <string name="title_toolbar_map">Местоположение</string>
